requirements.txt
python-dotenv~=1.0.0
streamlit==1.30.0
langchain==0.0.339
openai==1.1.0
tiktoken==0.5.1
boto3==1.28.84
nougat-ocr==0.1.17
tenacity==8.2.2
requests~=2.31.0
